requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
deepface==0.0.79
google-generativeai==0.3.2
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
pillow==10.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-multipart==0.0.6